% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvvarDATA.R
\name{tvvarDATA}
\alias{tvvarDATA}
\title{xxx}
\usage{
tvvarDATA(
  data,
  tvvarOpt = "TVVAR",
  nb = 10,
  pbar = TRUE,
  scale = FALSE,
  consec = NULL
)
}
\arguments{
\item{data}{An \eqn{(nt\times nv)}{(nt x nv)} data matrix.}

\item{tvvarOpt}{xxx (default = \code{"TVVAR"}).}

\item{nb}{xxx (default = 10).}

\item{pbar}{xxx (default = \code{TRUE}).}

\item{scale}{xxx (default = \code{FALSE}).}

\item{consec}{xxx (default = \code{NULL}).}
}
\value{
The function returns a list with 1 element:
  \item{model}{xxx.}
}
\description{
\code{tvvarDATA} xxx.
}
\section{Details}{
 xxx.
}

\examples{
# Example 1 - xxx

# Example 2 - xxx
}
\references{
\insertRef{RobertsLaughlin1996}{GGUM}

\insertRef{Robertsetal2000}{GGUM}
}
\author{
Laura Bringmann, \email{l.f.bringmann@rug.nl}
}
