% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{tvvarGAM-package}
\alias{tvvarGAM-package}
\alias{_PACKAGE}
\title{tvvarGAM}
\description{
With this package one can estimate time-varying vector
autoregressive models based on generalized additive models. It uses a
regression form, making it easy to use. At the moment it can only handle
gradual smooth change of both the intercept and the vector autoregressive
parameters.

Laura and Jonas: This is the front page of the package on the Help tab.
Please provide some details, including formulas and refs if you will,
laying out the contents of the package.

Versions:
\itemize{
\item Version 0.2.0 (April 2020)\cr
The code was overall optimized. Plot and summary methods for objects of
class 'tvvarGAM' were created.
\item Version 0.1.1 (January 2019 ?????)\cr
Laura: Briefly mention what this version of the package offered.
}
}
\details{
\tabular{ll}{
Package: \tab tvvarGAM\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 2020-04-04\cr
License: \tab GPL Version 2 or later\cr }

The \pkg{tvvarGAM} package contains useful functions, summarized below:
\itemize{
\item \code{\link{tvvarGAM}}: xxx
\item \code{\link{tvvarDATA}}: xxx
\item \code{\link{tvvarSIM}}: xxx
\item Available methods for objects of class "GGUM":
\tabular{l}{
plot() \cr
print() \cr
summary() \cr
} }
}
\examples{
# Example 1 - xxx

# Example 2 - xxx

}
\references{
\insertRef{RobertsLaughlin1996}{GGUM}

\insertRef{Robertsetal2000}{GGUM}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LauraBringmann/tvvarGAM}
}

}
\author{
\strong{Maintainer}: Laura Bringmann \email{l.f.bringmann@rug.nl}

Authors:
\itemize{
  \item Jonas Haslbeck \email{jonashaslbeck@gmail.com}
  \item Jorge N. Tendeiro \email{j.n.tendeiro@rug.nl}
}

}
