% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvvarGAM.R
\name{tvvarGAM}
\alias{tvvarGAM}
\title{Fit the xxx model (gam)}
\usage{
tvvarGAM(
  data = NULL,
  nb = 10,
  consec = NULL,
  scale = FALSE,
  beepvar = NULL,
  dayvar = NULL,
  tvvarOpt = "TVVAR",
  thresholding = FALSE,
  pbar = TRUE
)
}
\arguments{
\item{data}{An \eqn{(nt\times nv)}{(nt x nv)} data matrix, or an object of
class 'tvvarSIM'.}

\item{nb}{xxx (default = 10).}

\item{consec}{xxx (default = \code{NULL}).}

\item{scale}{xxx (default = \code{FALSE}).}

\item{beepvar}{xxx (default = \code{NULL}).}

\item{dayvar}{xxx (default = \code{NULL}).}

\item{tvvarOpt}{xxx (default = \code{"TVVAR"}).}

\item{thresholding}{xxx (default = \code{FALSE}).}

\item{pbar}{xxx (default = \code{TRUE}).}
}
\value{
The function returns a list (an object of class \code{tvvarGAM}) with 3
  elements:
  \item{call}{xxx.}
  \item{Results_GAM}{xxx.}
  \item{model}{xxx.}
}
\description{
\code{tvvarGAM} xxx.
}
\section{Details}{
 xxx.
}

\examples{
# Example 1 - xxx

# Example 2 - xxx
}
\references{
\insertRef{RobertsLaughlin1996}{GGUM}

\insertRef{Robertsetal2000}{GGUM}
}
\author{
Laura Bringmann, \email{l.f.bringmann@rug.nl}
}
