% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvvarSimulate.R
\name{tvvarSIM}
\alias{tvvarSIM}
\title{xxx}
\usage{
tvvarSIM(
  nt,
  nv,
  FUN.aint = rep(1, nv),
  max.aint = rep(1, nv),
  FUN.rho = rep(1, nv * nv),
  max.rho = rep(0.4, nv * nv),
  min.cor = 0.1,
  max.cor = 0.5
)
}
\arguments{
\item{nt}{Number of time points.}

\item{nv}{Number of variables.}

\item{FUN.aint}{Vector of length \eqn{nv}{nv} with 1s=invariant, 2s=linear,
3s=sine. Default is a vector of 1s.}

\item{max.aint}{Vector of length \eqn{nv}{nv} with maximum value for the
intercept. Default is a vector of 1s.}

\item{FUN.rho}{Vector of length \eqn{nv^2}{nv^2} with 1s=invariant,
2s=linear, 3s=sine. Default is a vector of 1s.}

\item{max.rho}{Vector of length \eqn{nv^2}{nv^2} with maximum value for the
off-diagonal correlations. Default is a vector of .4s.}

\item{min.cor}{Minimum correlation value. Default is .1.}

\item{max.cor}{Maximum correlation value. Default is .5.}
}
\value{
The function returns a list (an object of class \code{tvvarSIM})
  with 4 elements:
  \item{y}{xxx.}
  \item{aint}{xxx.}
  \item{rho}{xxx.}
  \item{sigma}{xxx.}
}
\description{
\code{tvvarSIM} xxx.
}
\section{Details}{
 xxx.
}

\examples{
# Example 1 - xxx

# Example 2 - xxx
}
\references{
\insertRef{RobertsLaughlin1996}{GGUM}

\insertRef{Robertsetal2000}{GGUM}
}
\author{
Laura Bringmann, \email{l.f.bringmann@rug.nl}
}
